Implements iPhone skin for iAnki.
The skin is implemented as an additional CSS file (ianki_ext/static/iphone.css)
that is included only on iPod and iPhone. In order to implement this, a few
changes to the HTML template and the Javascript were necessary. More notably,
the "Show Answer" buttom had to be move below the div used for the answer. A
few additional entries also needed to be added to base.css and are overriden by
iphone.css.